requirements.txt
discord.py
openai
pyyaml
python-dotenv
